Replace can_network -> can_router
hal::can_router is a much simplier version of can_network and doesn't
force driver developers to use polling. hal::can_router maps can message
IDs to callbacks.
