--- tests/can/network.test.cpp
@@ -1,1 +1,0 @@
-#include <libhal/can/network.hpp>
