+++ tests/can/router.test.cpp
@@ -1,0 +1,1 @@
+#include <libhal/can/router.hpp>
